refactor(mint): split FCT and FIT workflows
