+++ crates/forge_pen_test/src/forge_pentester.rs
@@ -4,15 +4,15 @@ use std::cell::RefCell;
 
 use forge_loader::manifest::ForgeManifest;
 use serde::Deserialize;
-use tracing::{debug, info, warn};
+use tracing::{info, warn};
 use ureq::Agent;
 use url::Url;
 
 use crate::app_config::{AppConfig, ExtensionConfig};
 use crate::fsrt_remote_config::FsrtRemoteConfig;
 use crate::mint_common::{
-    GraphqlErrorObject, GraphqlHeaders, GraphqlRequest, JwtValidity, PenTestError,
-    build_cookie_header, decode_jwt_payload, fetch_cloud_id, post_graphql,
+    GraphqlHeaders, GraphqlRequest, JwtValidity, PenTestError, build_cookie_header,
+    decode_jwt_payload, fetch_cloud_id, post_graphql,
 };
 
 const PRODUCTION_ENVIRONMENT_KEY: &str = "production";
@@ -60,28 +60,6 @@ const APPS_QUERY: &str = r#"query Apps(
     }
   }
 }"#;
-
-const DEFAULT_GLOBAL_APP_MUTATION: &str = r#"mutation SignForgeContextToken($input: GlobalAppSignForgeContextTokensInput!) {
-  globalApp_signForgeContextTokens(input: $input) {
-    success
-    errors {
-      message
-    }
-    tokens {
-      jwt
-    }
-  }
-}"#;
-const GLOBAL_APP_OPERATION_NAME: &str = "SignForgeContextToken";
-const FIT_MUTATION: &str = r#"mutation SignInvocationTokenForUI($input: SignInvocationTokenForUIInput!) {
-  signInvocationTokenForUI(input: $input) {
-    forgeInvocationToken {
-      jwt
-    }
-  }
-}"#;
-const FIT_OPERATION_NAME: &str = "SignInvocationTokenForUI";
-const REDACTED_FCT: &str = "<FCT selected at runtime>";
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 enum EnvironmentSelection {
@@ -157,42 +135,7 @@ struct DeployedExtension {
     extension_type_key: String,
 }
 
-#[derive(Debug, Deserialize)]
-struct FctData {
-    #[serde(rename = "globalApp_signForgeContextTokens")]
-    result: FctResult,
-}
-
-#[derive(Debug, Deserialize)]
-struct FctResult {
-    success: bool,
-    errors: Vec<GraphqlErrorObject>,
-    tokens: Vec<ForgeContextToken>,
-}
-
-#[derive(Debug, Deserialize)]
-struct ForgeContextToken {
-    jwt: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct FitData {
-    #[serde(rename = "signInvocationTokenForUI")]
-    result: Option<FitResult>,
-}
-
-#[derive(Debug, Deserialize)]
-struct FitResult {
-    #[serde(rename = "forgeInvocationToken")]
-    token: Option<ForgeInvocationToken>,
-}
-
-#[derive(Debug, Deserialize)]
-struct ForgeInvocationToken {
-    jwt: String,
-}
-
-/// Reusable client for minting Forge Context and Invocation Tokens.
+/// Reusable client for Forge penetration-testing operations.
 ///
 /// Reuses authentication, one [`Agent`], and a deployment snapshot.
 pub struct ForgePenTester<'manifest, 'source> {
@@ -291,39 +234,10 @@ impl<'manifest, 'source> ForgePenTester<'manifest, 'source> {
         }
     }
 
-    /// Constructs the exact GraphQL request that `mint_fct` will send.
-    pub fn mint_fct_request(
-        &self,
-        module_key: &str,
-        ctx: &serde_json::Value,
-    ) -> Result<GraphqlRequest<serde_json::Value>, PenTestError> {
-        if !ctx.is_object() {
-            return Err(PenTestError::InvalidFctContext);
-        }
-        let extension = self.config.extension_for_module_key(module_key)?;
-        info!(
-            module_key = ?module_key,
-            deployment_id = ?extension.deployment_id(),
-            extension_id = ?extension.extension_id(),
-            extension_type = ?extension.extension_type(),
-            "selected deployed extension"
-        );
-        Ok(GraphqlRequest {
-            operation_name: GLOBAL_APP_OPERATION_NAME.to_string(),
-            query: DEFAULT_GLOBAL_APP_MUTATION.to_string(),
-            variables: serde_json::json!({
-                "input": {
-                    "contextIds": [self.config.context_id()],
-                    "extensionContexts": [{
-                        "appVersion": self.config.app_version(),
-                        "context": ctx,
-                        "extensionId": extension.extension_id(),
-                        "extensionType": extension.extension_type(),
-                        "installationId": self.config.installation_id()
-                    }]
-                }
-            }),
-        })
+    pub(crate) fn replace_cached_fct_jwt(&self, jwt: &str) {
+        let mut cached_fct_jwt = self.cached_fct_jwt.borrow_mut();
+        cached_fct_jwt.clear();
+        cached_fct_jwt.push_str(jwt);
     }
 
     /// Sends a GraphQL mutation and returns its required, generic data object.
@@ -341,83 +255,6 @@ impl<'manifest, 'source> ForgePenTester<'manifest, 'source> {
         let mut graphql_endpoint = self.site.clone();
         graphql_endpoint.set_path("/gateway/api/graphql");
         post_graphql(&self.agent, &graphql_endpoint, request)
-    }
-
-    /// Mints a new FCT and replaces the cache only on success.
-    pub fn mint_fct(
-        &self,
-        module_key: &str,
-        ctx: &serde_json::Value,
-    ) -> Result<String, PenTestError> {
-        let request = self.mint_fct_request(module_key, ctx)?;
-        let variables = format!("{:#}", request.variables);
-        info!(variables = %variables, "FCT GraphQL variables");
-        let data: FctData = self.post_graphql_mutation(&request)?;
-        let FctResult {
-            success,
-            errors,
-            tokens,
-        } = data.result;
-        if !success || !errors.is_empty() {
-            return Err(PenTestError::MintRejected { success, errors });
-        }
-
-        let jwt = tokens
-            .into_iter()
-            .next()
-            .map(|token| token.jwt)
-            .ok_or(PenTestError::MissingFctToken)?;
-        self.cached_fct_jwt.borrow_mut().clone_from(&jwt);
-        debug!("successfully minted Forge Context Token");
-        Ok(jwt)
-    }
-
-    /// Constructs the exact GraphQL request used to mint a FIT.
-    pub fn mint_fit_request(
-        &self,
-        fct: &str,
-        remote_key: &str,
-    ) -> Result<GraphqlRequest<serde_json::Value>, PenTestError> {
-        if fct.trim().is_empty() {
-            return Err(PenTestError::EmptySuppliedFct);
-        }
-
-        Ok(GraphqlRequest {
-            operation_name: FIT_OPERATION_NAME.to_string(),
-            query: FIT_MUTATION.to_string(),
-            variables: serde_json::json!({
-                "input": {
-                    "forgeContextToken": fct,
-                    "remoteKey": remote_key
-                }
-            }),
-        })
-    }
-
-    /// Mints a FIT for a Forge remote using the supplied FCT.
-    pub fn mint_fit(&self, remote_key: &str, fct: &str) -> Result<String, PenTestError> {
-        let request = self.mint_fit_request(fct, remote_key)?;
-        let mut variables = request.variables.clone();
-        variables["input"]["forgeContextToken"] = REDACTED_FCT.into();
-        let variables = format!("{variables:#}");
-        info!(
-            operation_name = FIT_OPERATION_NAME,
-            variables = %variables,
-            "FIT GraphQL variables"
-        );
-        let token = (|| {
-            let data: FitData = self.post_graphql_mutation(&request)?;
-            data.result
-                .and_then(|result| result.token)
-                .map(|token| token.jwt)
-                .ok_or(PenTestError::MissingFitToken)
-        })()
-        .map_err(|source| PenTestError::FitMintFailed {
-            source: Box::new(source),
-            available: self.config.remote_keys(),
-        })?;
-        debug!("successfully minted Forge Invocation Token");
-        Ok(token)
     }
 }
 
@@ -674,8 +511,12 @@ mod tests {
             .mint_fct_request("installation-1-module", &context)
             .unwrap();
 
-        assert_eq!(request.operation_name, GLOBAL_APP_OPERATION_NAME);
-        assert_eq!(request.query, DEFAULT_GLOBAL_APP_MUTATION);
+        assert_eq!(request.operation_name, "SignForgeContextToken");
+        assert!(
+            request
+                .query
+                .contains("globalApp_signForgeContextTokens(input: $input)")
+        );
         assert_eq!(
             request.variables,
             serde_json::json!({
@@ -752,8 +593,12 @@ mod tests {
             .mint_fit_request(" fct ", "not-declared-in-manifest")
             .unwrap();
 
-        assert_eq!(request.operation_name, FIT_OPERATION_NAME);
-        assert_eq!(request.query, FIT_MUTATION);
+        assert_eq!(request.operation_name, "SignInvocationTokenForUI");
+        assert!(
+            request
+                .query
+                .contains("signInvocationTokenForUI(input: $input)")
+        );
         assert_eq!(
             request.variables,
             serde_json::json!({
